refactored package for swagger-generator.
