+++ modules/swagger-generator/src/main/java/io/swagger/v3/generator/exception/NotFoundException.java
@@ -12,18 +12,14 @@
  * the License.
  */
 
-package io.swagger.generator.exception;
+package io.swagger.v3.generator.exception;
 
-public class ApiException extends Exception {
-    private static final long serialVersionUID = -5085112752305370687L;
-    private int code;
+import org.apache.http.HttpStatus;
 
-    public ApiException(int code, String msg) {
-        super(msg);
-        this.code = code;
-    }
+public class NotFoundException extends ApiException {
+    private static final long serialVersionUID = -1223255119112336573L;
 
-    public int getCode() {
-        return code;
+    public NotFoundException(String msg) {
+        super(HttpStatus.SC_NOT_FOUND, msg);
     }
 }
