+++ modules/swagger-generator/src/main/java/io/swagger/v3/generator/model/GenerationRequest.java
@@ -1,4 +1,4 @@
-    package io.swagger.generator.model;
+    package io.swagger.v3.generator.model;
 
 import java.io.Serializable;
 import java.util.Map;
