+++ modules/swagger-generator/src/main/java/io/swagger/v3/generator/online/GeneratorController.java
@@ -1,4 +1,4 @@
-package io.swagger.generator.online;
+package io.swagger.v3.generator.online;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.swagger.codegen.v3.ClientOptInput;
@@ -6,9 +6,9 @@ import io.swagger.codegen.v3.CodegenConfig;
 import io.swagger.codegen.v3.CodegenType;
 import io.swagger.codegen.v3.DefaultGenerator;
 import io.swagger.codegen.v3.config.CodegenConfigurator;
-import io.swagger.generator.model.GenerationRequest;
-import io.swagger.generator.util.GeneratorUtil;
-import io.swagger.generator.util.ZipUtil;
+import io.swagger.v3.generator.model.GenerationRequest;
+import io.swagger.v3.generator.util.GeneratorUtil;
+import io.swagger.v3.generator.util.ZipUtil;
 import io.swagger.oas.inflector.models.RequestContext;
 import io.swagger.oas.inflector.models.ResponseContext;
 import io.swagger.v3.core.util.Json;
