+++ modules/swagger-generator/src/main/java/io/swagger/v3/generator/online/GeneratorServlet.java
@@ -1,4 +1,4 @@
-package io.swagger.generator.online;
+package io.swagger.v3.generator.online;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
