+++ modules/swagger-generator/src/main/java/io/swagger/v3/generator/util/ValidationMessage.java
@@ -1,4 +1,4 @@
-package io.swagger.generator.util;
+package io.swagger.v3.generator.util;
 
 public class ValidationMessage {
     private String path, message, severity;
